Potion damage buff works

+++ objects/enemy/bat.py
@@ -139,7 +139,7 @@ class Bat(Enemy):
         self.projectiles.append(warnProjectile)
         self.commitedToAttack = True
     
-    def onHit(self, damage: int, attack_id: int) -> None:
+    def onHit(self, damage: float, attack_id: int) -> None:
         super().onHit(damage, attack_id)
     
     def spawnCoins(self):
